requirements.txt
# Core Dependencies
pydantic==2.5.0
pydantic-settings==2.1.0
pyyaml==6.0.1
fastapi==0.104.1
uvicorn==0.24.0
nest-asyncio==1.5.8

# Document Processing
# PyPDF2==3.0.1  # Removed - using pdfplumber for all PDF operations
pdfplumber==0.10.3  # For PDF text and table extraction with layout preservation
python-docx==0.8.11
openpyxl==3.1.2
python-pptx==0.6.23
tiktoken==0.5.2
Pillow==10.1.0

# Knowledge Graph
SPARQLWrapper==2.0.0
rdflib==7.0.0

# Vector Store & RAG
chromadb==0.4.22
sentence-transformers==2.2.2

# LLM Clients
requests==2.31.0
httpx==0.25.2
openai==1.12.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
ruff==0.1.9
black==23.12.1